requirements.txt
argcomplete==3.5.1
attrs==24.2.0
Automat==24.8.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
boto3==1.35.69
botocore==1.35.69
bottle==0.12.25
Brlapi==0.8.5
cached-property==1.5.2
cairocffi==1.7.0
CairoSVG==2.7.0
catfish==4.18.0
certifi==2023.5.7
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
cssselect2==0.7.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
defusedxml==0.7.1
distro==1.9.0
dnf==4.22.0
docker==7.1.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
file-magic==0.4.0
filelock==3.16.1
Grammalecte-fr==2.1.2
h11==0.14.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.11.2
langtable==0.0.68
legacy-cgi==2.6
libcomps==0.1.21
libdnf==0.73.4
louis==3.28.0
lxml==5.2.1
nftables==0.1
ntpsec==1.2.3
numpy==2.1.3
olefile==0.47
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
parsel==1.9.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
plotly==5.24.1
plotly-express==0.4.1
ply==3.11
Protego==0.3.1
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycairo==1.25.1
pycparser==2.20
pycups==2.0.4
PyDispatcher==2.0.7
pyenchant==3.2.2
PyGObject==3.48.2
PyNaCl==1.5.0
pyOpenSSL==24.3.0
PySocks==1.7.1
pystray==0.17.3
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-gettext==4.0
python-manatools==0.0.4
python-xlib==0.33
pytz==2024.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
queuelib==1.7.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rpds-py==0.19.1
rpm==4.20.0
s3transfer==0.10.4
scipy==1.14.1
Scrapy==2.12.0
scrapy-selenium==0.0.7
selenium==4.27.1
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
service-identity==24.2.0
setools==4.5.1
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.35-build/setroubleshoot-3.3.35/src
setuptools==69.2.0
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.8.1
soupsieve==2.6
statsmodels==0.14.4
systemd-python==235
tenacity==9.0.0
texttable==1.6.7
tinycss2==1.3.0
tldextract==5.1.3
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.10.0
typing_extensions==4.12.2
tzdata==2024.2
ufw==0.35
urllib3==1.26.20
w3lib==2.2.1
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
